+++ data/tasks.txt
@@ -1,3 +1,4 @@
 
 1. [T][ ] task1
 2. [D][X] deadline task2 (by: Aug 21 2022 06:00)
+3. [T][ ] task3
