+++ src/main/java/duke/CommandWords.java
@@ -1,3 +1,5 @@
+package duke;
+
 public enum CommandWords {
     BYE, LIST, UNMARK, MARK, TODO, EVENT, DEADLINE, DELETE
 }
