+++ src/main/java/duke/Ui.java
@@ -1,3 +1,6 @@
+package duke;
+
+import duke.task.Task;
 import java.util.Scanner;
 
 public class Ui {
