+++ src/main/java/duke/command/AddCommand.java
@@ -1,3 +1,11 @@
+package duke.command;
+
+import duke.DukeException;
+import duke.TaskList;
+import duke.Ui;
+import duke.Storage;
+import duke.task.Task;
+
 public class AddCommand extends Command {
 
     Task task;
