+++ src/main/java/duke/command/Command.java
@@ -1,3 +1,10 @@
+package duke.command;
+
+import duke.DukeException;
+import duke.TaskList;
+import duke.Ui;
+import duke.Storage;
+
 abstract public class Command {
 
     public abstract void execute(TaskList tasks, Ui ui, Storage storage) throws DukeException;
