+++ src/main/java/duke/task/Task.java
@@ -1,3 +1,5 @@
+package duke.task;
+
 abstract public class Task {
     private String description;
     private boolean isDone;
